day 18/requirements.txt
fastapi
uvicorn
python-dotenv
requests
assemblyai
Jinja2
python-multipart